refactor(provenance): hoist generic claim helpers out of catalog
Move the model-agnostic claim-construction helpers from catalog into
provenance, the layer that owns Claim and the relationship-schema
registry, so the ingest front end can reach them without importing
catalog:

- new apps/provenance/claims.py: build_relationship_claim, the three
  normalizers (fk / abbreviation / alias) and AliasIdentity. RelationshipClaim
  is promoted from a bare tuple alias to a NamedTuple(claim_key, value) —
  a tuple subclass, so every unpack site is untouched.
- apps/catalog/resolve/claim_presence.py -> apps/provenance/claim_presence.py
  (member_is_present + PresenceClaim), unchanged.
- get_relationship_namespaces: catalog's one-line re-export dropped;
  importers now hit provenance.validation directly.

catalog/claims.py keeps only its catalog-coupled residue
(register_catalog_relationship_schemas, build_media_attachment_claim,
make_authoritative_scope, get_all_namespace_keys).

The exists-filter guard documented validation.py's inline spelling as
temporary "until the leaf moves down to a shared layer" — now satisfied:
validation.py routes through member_is_present, the guard's allowlist
collapses to the single canonical file, and the guard moves to
provenance/tests/ beside the ranking guard it parallels.

Clears 6 of the 11 ingest->catalog baseline edges; lint-imports stays
green with 5 remaining (CatalogModel x3, resolve, resolve._entities) for
the later steps.
