+++ backend/apps/provenance/tests/test_exists_filter_is_canonical.py
@@ -1,32 +1,26 @@
 """Guard: the relationship-member exists-filter lives in exactly one place.
 
 A member claim's ``exists`` flag (``exists=false`` is a tombstone) is read
-through :func:`apps.catalog.resolve.claim_presence.member_is_present`.  This test
+through :func:`apps.provenance.claim_presence.member_is_present`.  This test
 fails if any other module re-spells the ``value.get("exists", True)`` filter
 inline instead of routing through the shared leaf — the same drift the ranking
 guard (``apps/provenance/tests/test_ranking_is_canonical.py``) prevents for the
 winner-pick.
 
-Two files are allow-listed:
-
-* ``claim_presence.py`` *defines* the leaf.
-* ``provenance/validation.py`` sits *below* catalog, so it cannot import the
-  catalog-located leaf; it keeps its own spelling until the leaf moves down to a
-  shared layer.
+Only ``claim_presence.py`` (which *defines* the leaf) is allow-listed.  The
+leaf now lives in provenance, below catalog, so every layer — including
+``provenance/validation.py`` — can and does route through it.
 """
 
 from __future__ import annotations
 
 from pathlib import Path
 
-import apps.catalog.resolve.claim_presence as presence
+import apps.provenance.claim_presence as presence
 
 _CANONICAL = Path(presence.__file__).resolve()
-_APPS_DIR = _CANONICAL.parents[2]  # .../backend/apps
-_ALLOWLISTED = {
-    _CANONICAL,
-    (_APPS_DIR / "provenance" / "validation.py").resolve(),
-}
+_APPS_DIR = _CANONICAL.parents[1]  # .../backend/apps
+_ALLOWLISTED = {_CANONICAL}
 
 
 def _production_sources() -> list[Path]:
@@ -47,5 +41,5 @@ def test_exists_filter_routed_through_member_is_present() -> None:
     ]
     assert not offenders, (
         "Inlined the member exists-filter instead of calling member_is_present: "
-        f"{offenders}. Route through apps.catalog.resolve.claim_presence."
+        f"{offenders}. Route through apps.provenance.claim_presence."
     )
